CMS Docker Teste Code Python/requirements.txt
schedule
mysql-connector-python
SQLAlchemy
protobuf
six